pyproject.toml
[tool.poetry]
name = "trading"
version = "0.1.0"
description = "Trading Strategy Platform with MA Cross, Portfolio Optimization, and GraphQL API"
authors = ["Cole Morton"]
readme = "README.md"
package-mode = true
packages = [
    {include = "app"}
]

[tool.poetry.scripts]
trading-cli = "app.cli.main:cli_main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
numpy = ">=1.16.5,<2.0.0"
yfinance = "^0.2.41"
seaborn = "^0.13.2"
polars = "^1.5.0"
flask = "^3.0.3"
jupyter = "^1.1.1"
notebook = "^7.2.2"
ipywidgets = "^8.1.5"
pybit = "^5.8.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
langchain-community = "^0.3.3"
langchain-experimental = "^0.3.2"
langchain-ollama = "^0.2.0"
tabulate = "^0.9.0"
skfolio = "^0.6.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
strawberry-graphql = {version = "^0.236.0", extras = ["fastapi"]}
psutil = "^7.0.0"
langchain-anthropic = "^0.3.13"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.0"
redis = "^5.0.1"
prisma = "^0.11.0"
pydantic-settings = "^2.9.1"
vectorbt = "^0.28.0"
pyarrow = "^20.0.0"
cachetools = "^6.1.0"
typer = "^0.9.0"
rich = "^13.0.0"
pyyaml = "^6.0.0"
kaleido = "0.2.1"
plotly = "5.24.1"
anywidget = "^0.9.0"
jsonschema = "^4.25.1"
jsonschema-specifications = "^2025.4.1"
arq = "^0.25.0"
bcrypt = "^4.0.0"
psycopg2-binary = "^2.9.0"
greenlet = "^3.0.0"
sqlmodel = "^0.0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"
black = "^24.0.0"
isort = "^5.13.0"
ruff = "^0.8.0"
flake8 = "^7.0.0"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^24.0.0"
mypy = "^1.8.0"
pylint = "^3.0.0"
bandit = "^1.7.0"
vulture = "^2.11"
pre-commit = "^3.6.0"
autoflake = "^2.3.1"
autopep8 = "^2.3.2"
types-requests = "^2.32.0.20250602"
pytest-mock = "^3.14.1"
pytest-timeout = "^2.4.0"
pytest-xdist = "^3.7.0"
hypothesis = "^6.135.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
# Make MyPy more practical for large codebases
ignore_errors = false
show_error_codes = true
# Allow untyped calls for now - we'll gradually improve
disallow_untyped_calls = false
# Focus on critical errors first - gradually enabling type checking
disable_error_code = ["misc", "union-attr", "attr-defined", "import-untyped", "unreachable", "index", "literal-required", "call-arg", "assignment", "arg-type", "name-defined", "method-assign"]
# Alternatively, use a more permissive approach for now
ignore_missing_imports = true

# Ignore third-party packages without type stubs
[[tool.mypy.overrides]]
module = [
    "vectorbt.*",
    "yfinance.*",
    "seaborn.*",
    "polars.*",
    "pybit.*",
    "langchain.*",
    "skfolio.*",
    "strawberry.*",
    "prisma.*",
    "plotly.*",
    "pandas_ta.*",
    "ta.*"
]
ignore_missing_imports = true

# Exclude vendor/external code
[[tool.mypy.overrides]]
module = "app.trading_bot.trendspider.*"
ignore_errors = true

[tool.pylint.messages_control]
disable = "C0103,C0114,C0115,C0116,R0903,R0913,W0613"

[tool.pylint.format]
max-line-length = 88

[tool.vulture]
min_confidence = 80
paths = ["app", "tests"]
exclude = ["*/migrations/*", "*/tests/*"]

[tool.ruff]
# Ruff configuration - modern, fast Python linter
line-length = 88
target-version = "py310"

# Exclude directories
extend-exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".tox",
    ".venv",
    "build",
    "dist",
    "__pycache__",
    "node_modules",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "RET",   # flake8-return
    "PTH",   # flake8-use-pathlib
    "TD",    # flake8-todos
    "FIX",   # flake8-fixme
    "ERA",   # eradicate (commented out code)
    "PL",    # pylint
    "RUF",   # ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",    # line too long (handled by black)
    "TD002",   # missing TODO author
    "TD003",   # missing TODO link
    "FIX002",  # line contains TODO
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "RET504",  # unnecessary variable assignment before return
    "SIM108",  # use ternary operator (sometimes less readable)
]

[tool.ruff.lint.per-file-ignores]
# Allow unused imports in __init__.py files
"__init__.py" = ["F401", "E402"]
# Relax some rules for tests
"tests/**/*.py" = ["PLR2004", "S101", "S105", "S106"]
# Relax rules for scripts
"scripts/**/*.py" = ["T201", "S101"]

[tool.ruff.lint.isort]
# Make isort compatible with black
combine-as-imports = true
force-sort-within-sections = true
known-first-party = ["app"]
lines-after-imports = 2

[tool.ruff.lint.mccabe]
# Maximum cyclomatic complexity
max-complexity = 10

[tool.ruff.format]
# Use Black-compatible formatting
quote-style = "double"
indent-style = "space"
line-ending = "auto"
